+++ src/Detach/Inline.cs
@@ -1,182 +1,85 @@
-using Detach.Numerics;
-using System.Numerics;
-
 namespace Detach;
 
 /// <summary>
-/// Unsafe methods to quickly format values into a <see cref="Span{T}"/> without allocating memory.
-/// These must only be used inline, as the <see cref="Span{T}"/> is only valid until the next method call.
+/// <para>
+/// Unsafe methods to quickly format values into either a fixed UTF-8 or a fixed UTF-16 buffer, without allocating heap memory.
+/// These must only be used inline, as the buffers are only valid until the next method call.
+/// </para>
+/// <para>
+/// The buffers have a fixed size of 2048 characters, so the total length of the formatted string cannot exceed this limit.
+/// </para>
 /// </summary>
-public static class Inline
+public static partial class Inline
 {
-	private static readonly char[] _buffer = new char[2048];
+	private const string _separatorUtf16 = ", ";
 
-	public static Span<char> Buffer => _buffer;
+	private static readonly byte[] _bufferUtf8 = new byte[2048];
+	private static readonly char[] _bufferUtf16 = new char[2048];
 
-	public static ReadOnlySpan<char> Span(InlineInterpolatedStringHandler interpolatedStringHandler)
+	private static ReadOnlySpan<byte> SeparatorUtf8 => ", "u8;
+
+	public static Span<byte> BufferUtf8 => _bufferUtf8;
+	public static Span<char> BufferUtf16 => _bufferUtf16;
+
+	public static ReadOnlySpan<byte> Utf8(InlineInterpolatedStringHandlerUtf8 interpolatedStringHandler)
 	{
 		return interpolatedStringHandler;
 	}
 
-	public static ReadOnlySpan<char> Span<T>(T t, ReadOnlySpan<char> format = default, IFormatProvider? provider = null)
-		where T : ISpanFormattable
+	public static ReadOnlySpan<byte> Utf8<T>(T t, ReadOnlySpan<char> format = default, IFormatProvider? provider = null)
+		where T : IUtf8SpanFormattable
 	{
-		return t.TryFormat(_buffer, out int charsWritten, format, provider) ? _buffer.AsSpan()[..charsWritten] : ReadOnlySpan<char>.Empty;
+		if (!t.TryFormat(BufferUtf8, out int charsWritten, format, provider))
+			throw new InvalidOperationException("The formatted string is too long.");
+
+		return BufferUtf8[..charsWritten];
 	}
 
-	public static ReadOnlySpan<char> Span(Vector2 value, ReadOnlySpan<char> format = default, IFormatProvider? provider = default)
+	private static void WriteUtf8(ref int charsWritten, ReadOnlySpan<byte> value)
 	{
-		int charsWritten = 0;
-		TryWrite(_buffer, ref charsWritten, value.X, format, provider);
-		TryWriteString(_buffer, ref charsWritten, ", ");
-		TryWrite(_buffer, ref charsWritten, value.Y, format, provider);
-		return _buffer.AsSpan(0, charsWritten);
-	}
+		if (!value.TryCopyTo(BufferUtf8[charsWritten..]))
+			throw new InvalidOperationException("The formatted string is too long.");
 
-	public static ReadOnlySpan<char> Span(Vector3 value, ReadOnlySpan<char> format = default, IFormatProvider? provider = default)
-	{
-		int charsWritten = 0;
-		TryWrite(_buffer, ref charsWritten, value.X, format, provider);
-		TryWriteString(_buffer, ref charsWritten, ", ");
-		TryWrite(_buffer, ref charsWritten, value.Y, format, provider);
-		TryWriteString(_buffer, ref charsWritten, ", ");
-		TryWrite(_buffer, ref charsWritten, value.Z, format, provider);
-		return _buffer.AsSpan(0, charsWritten);
-	}
-
-	public static ReadOnlySpan<char> Span(Vector4 value, ReadOnlySpan<char> format = default, IFormatProvider? provider = default)
-	{
-		int charsWritten = 0;
-		TryWrite(_buffer, ref charsWritten, value.X, format, provider);
-		TryWriteString(_buffer, ref charsWritten, ", ");
-		TryWrite(_buffer, ref charsWritten, value.Y, format, provider);
-		TryWriteString(_buffer, ref charsWritten, ", ");
-		TryWrite(_buffer, ref charsWritten, value.Z, format, provider);
-		TryWriteString(_buffer, ref charsWritten, ", ");
-		TryWrite(_buffer, ref charsWritten, value.W, format, provider);
-		return _buffer.AsSpan(0, charsWritten);
-	}
-
-	public static ReadOnlySpan<char> Span(Quaternion value, ReadOnlySpan<char> format = default, IFormatProvider? provider = default)
-	{
-		int charsWritten = 0;
-		TryWrite(_buffer, ref charsWritten, value.X, format, provider);
-		TryWriteString(_buffer, ref charsWritten, ", ");
-		TryWrite(_buffer, ref charsWritten, value.Y, format, provider);
-		TryWriteString(_buffer, ref charsWritten, ", ");
-		TryWrite(_buffer, ref charsWritten, value.Z, format, provider);
-		TryWriteString(_buffer, ref charsWritten, ", ");
-		TryWrite(_buffer, ref charsWritten, value.W, format, provider);
-		return _buffer.AsSpan(0, charsWritten);
-	}
-
-	public static ReadOnlySpan<char> Span(Rgba value, ReadOnlySpan<char> format = default, IFormatProvider? provider = default)
-	{
-		int charsWritten = 0;
-		TryWrite(_buffer, ref charsWritten, value.R, format, provider);
-		TryWriteString(_buffer, ref charsWritten, ", ");
-		TryWrite(_buffer, ref charsWritten, value.G, format, provider);
-		TryWriteString(_buffer, ref charsWritten, ", ");
-		TryWrite(_buffer, ref charsWritten, value.B, format, provider);
-		TryWriteString(_buffer, ref charsWritten, ", ");
-		TryWrite(_buffer, ref charsWritten, value.A, format, provider);
-		return _buffer.AsSpan(0, charsWritten);
-	}
-
-	public static ReadOnlySpan<char> Span(Matrix3x2 value, ReadOnlySpan<char> format = default, IFormatProvider? provider = default)
-	{
-		int charsWritten = 0;
-		TryWrite(_buffer, ref charsWritten, value.M11, format, provider);
-		TryWriteString(_buffer, ref charsWritten, ", ");
-		TryWrite(_buffer, ref charsWritten, value.M12, format, provider);
-		TryWriteString(_buffer, ref charsWritten, ", ");
-		TryWrite(_buffer, ref charsWritten, value.M21, format, provider);
-		TryWriteString(_buffer, ref charsWritten, ", ");
-		TryWrite(_buffer, ref charsWritten, value.M22, format, provider);
-		TryWriteString(_buffer, ref charsWritten, ", ");
-		TryWrite(_buffer, ref charsWritten, value.M31, format, provider);
-		TryWriteString(_buffer, ref charsWritten, ", ");
-		TryWrite(_buffer, ref charsWritten, value.M32, format, provider);
-		return _buffer.AsSpan(0, charsWritten);
-	}
-
-	public static ReadOnlySpan<char> Span(Matrix4x4 value, ReadOnlySpan<char> format = default, IFormatProvider? provider = default)
-	{
-		int charsWritten = 0;
-		TryWrite(_buffer, ref charsWritten, value.M11, format, provider);
-		TryWriteString(_buffer, ref charsWritten, ", ");
-		TryWrite(_buffer, ref charsWritten, value.M12, format, provider);
-		TryWriteString(_buffer, ref charsWritten, ", ");
-		TryWrite(_buffer, ref charsWritten, value.M13, format, provider);
-		TryWriteString(_buffer, ref charsWritten, ", ");
-		TryWrite(_buffer, ref charsWritten, value.M14, format, provider);
-		TryWriteString(_buffer, ref charsWritten, ", ");
-		TryWrite(_buffer, ref charsWritten, value.M21, format, provider);
-		TryWriteString(_buffer, ref charsWritten, ", ");
-		TryWrite(_buffer, ref charsWritten, value.M22, format, provider);
-		TryWriteString(_buffer, ref charsWritten, ", ");
-		TryWrite(_buffer, ref charsWritten, value.M23, format, provider);
-		TryWriteString(_buffer, ref charsWritten, ", ");
-		TryWrite(_buffer, ref charsWritten, value.M24, format, provider);
-		TryWriteString(_buffer, ref charsWritten, ", ");
-		TryWrite(_buffer, ref charsWritten, value.M31, format, provider);
-		TryWriteString(_buffer, ref charsWritten, ", ");
-		TryWrite(_buffer, ref charsWritten, value.M32, format, provider);
-		TryWriteString(_buffer, ref charsWritten, ", ");
-		TryWrite(_buffer, ref charsWritten, value.M33, format, provider);
-		TryWriteString(_buffer, ref charsWritten, ", ");
-		TryWrite(_buffer, ref charsWritten, value.M34, format, provider);
-		TryWriteString(_buffer, ref charsWritten, ", ");
-		TryWrite(_buffer, ref charsWritten, value.M41, format, provider);
-		TryWriteString(_buffer, ref charsWritten, ", ");
-		TryWrite(_buffer, ref charsWritten, value.M42, format, provider);
-		TryWriteString(_buffer, ref charsWritten, ", ");
-		TryWrite(_buffer, ref charsWritten, value.M43, format, provider);
-		TryWriteString(_buffer, ref charsWritten, ", ");
-		TryWrite(_buffer, ref charsWritten, value.M44, format, provider);
-		return _buffer.AsSpan(0, charsWritten);
-	}
-
-	public static ReadOnlySpan<char> Span(string str)
-	{
-		return str.AsSpan();
-	}
-
-	private static bool TryWriteChar(Span<char> destination, ref int charsWritten, char value)
-	{
-		if (destination.IsEmpty)
-			return false;
-
-		if (charsWritten + 1 >= destination.Length)
-			return false;
-
-		destination[charsWritten++] = value;
-		return true;
-	}
-
-	private static bool TryWriteString(Span<char> destination, ref int charsWritten, string value)
-	{
-		if (destination.IsEmpty)
-			return false;
-
-		if (charsWritten + value.Length >= destination.Length)
-			return false;
-
-		value.AsSpan().CopyTo(destination[charsWritten..]);
 		charsWritten += value.Length;
-		return true;
 	}
 
-	private static bool TryWrite<T>(Span<char> destination, ref int charsWritten, T value, ReadOnlySpan<char> format = default, IFormatProvider? provider = null)
-		where T : ISpanFormattable
+	private static void WriteUtf8<T>(ref int charsWritten, T value, ReadOnlySpan<char> format = default, IFormatProvider? provider = null)
+		where T : IUtf8SpanFormattable
 	{
-		if (destination.IsEmpty)
-			return false;
-
-		if (!value.TryFormat(destination[charsWritten..], out int charsWrittenValue, format, provider))
-			return false;
+		if (!value.TryFormat(BufferUtf8[charsWritten..], out int charsWrittenValue, format, provider))
+			throw new InvalidOperationException("The formatted string is too long.");
 
 		charsWritten += charsWrittenValue;
-		return true;
+	}
+
+	public static ReadOnlySpan<char> Utf16(InlineInterpolatedStringHandlerUtf16 handler)
+	{
+		return handler;
+	}
+
+	public static ReadOnlySpan<char> Utf16<T>(T t, ReadOnlySpan<char> format = default, IFormatProvider? provider = null)
+		where T : ISpanFormattable
+	{
+		if (!t.TryFormat(_bufferUtf16, out int charsWritten, format, provider))
+			throw new InvalidOperationException("The formatted string is too long.");
+
+		return BufferUtf16[..charsWritten];
+	}
+
+	private static void WriteUtf16(ref int charsWritten, string value)
+	{
+		if (!value.AsSpan().TryCopyTo(BufferUtf16[charsWritten..]))
+			throw new InvalidOperationException("The formatted string is too long.");
+
+		charsWritten += value.Length;
+	}
+
+	private static void WriteUtf16<T>(ref int charsWritten, T value, ReadOnlySpan<char> format = default, IFormatProvider? provider = null)
+		where T : ISpanFormattable
+	{
+		if (!value.TryFormat(BufferUtf16[charsWritten..], out int charsWrittenValue, format, provider))
+			throw new InvalidOperationException("The formatted string is too long.");
+
+		charsWritten += charsWrittenValue;
 	}
 }